% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{type_alias}
\alias{type_alias}
\title{Create an alias for a defined type}
\usage{
type_alias(name)
}
\arguments{
\item{name}{Type to alias.
This can be a string, S3 \code{generic.class} or \code{class}. See examples}
}
\value{
A closure that retrieves the aliased type
}
\description{
Create an alias for a defined type
}
\examples{
library(types)
type.numeric <- type_define(is.numeric)

# Can apply name in four ways
type.num <- type_alias(type.numeric)
type.num <- type_alias(numeric)
type.num <- type_alias("type.numeric")
type.num <- type_alias("numeric")

f <- function(x ? num) x
ff <- type_check(f)
f(5)
\dontrun{
f(TRUE) # Error: `x` is a `logical` not a `numeric`.
f("a") # Error: `x` is a `character` not a `numeric`.
}

type.data.frame <- define_type(is.data.frame)
type.df <- type_alias(data.frame)

f <- function(x = ? df){ x } ? data.frame
ff <- type_check(f)

ff(data.frame(x = 1:10, y = 10:1))
\dontrun{
ff(1:10) # Error: `x` is a `integer` not a `data.frame`.
}
}

