% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{type_check_package}
\alias{type_check_package}
\title{Add type checking to all functions in a package}
\usage{
type_check_package(env = parent.frame(), ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{type_check}}}
}
\description{
This function can either be placed at the end of the collation order, or as
a drop-in for the \code{\link{.onLoad}} function.
}
\examples{
\dontrun{
.onLoad <- typeCheck::type_check_package

}
}

