% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{type_get}
\alias{type_get}
\title{Retrieve a given type from the type registry}
\usage{
type_get(name)
}
\arguments{
\item{name}{Type name to retrieve}
}
\value{
A type object if defined, or \code{NULL}
}
\description{
Retrieve a given type from the type registry
}

