% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{type}
\alias{type}
\title{Retrieve a given type if defined}
\usage{
type(name)
}
\arguments{
\item{name}{Type name to retrieve}
}
\value{
A type object
}
\description{
The default method signals an error
}

