% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{type_check_package}
\alias{type_check_package}
\title{Add type checking to all functions in a package}
\usage{
type_check_package(env = parent.frame(), ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{type_check}}}
}
\description{
This function can either be placed at the end of the collation order, or as
a drop-in for the \code{\link{.onLoad}} function.
}
\examples{
\dontrun{
# Simply put the following after all function definitions in package to add
checks on package installation.
typeCheck::type_check_package(asNamespace(pkgname))

# Alternatively can add the checks when the package is loaded.
.onLoad <- function(libname, pkgname) {
  typeCheck::type_check_package(asNamespace(pkgname))
}
}
}

