% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{type_check_package}
\alias{type_check_package}
\title{Add type checking to all functions in a package}
\usage{
type_check_package(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{type_check}}}
}
\description{
This function can be placed anywhere in the package code and will be
installed just before the lazy load database is populated. The idea for this
function and the implementation implementation borrows heavily from
\href{https://github.com/gaborcsardi/argufy/blob/bceef7904eef178c9aa67709940a29c1483c3c13/R/transform.R#L115-L142}{argufy::argufy_me()}.
}
\examples{
\dontrun{
typeCheck::type_check_package()
}
}

