% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{type_check_package}
\alias{type_check_package}
\title{Add type checking to all functions in a package}
\usage{
type_check_package(libname, pkgname, ...)
}
\arguments{
\item{libname}{a character string giving the library directory where
    the package defining the namespace was found.}

\item{pkgname}{a character string giving the name of the package.}

\item{...}{Additional arguments passed to \code{\link{type_check}}}
}
\description{
This function can either be placed at the end of the collation order, or in
the \code{\link{.onLoad}} function.
}

